backend/requirements.txt
fastapi
uvicorn 
python-dotenv
requests
sqlalchemy
python-jose[cryptography]
bcrypt==4.1.2
passlib==1.7.4
asyncpg
pydantic[email]
psycopg2-binary
gitpython
sentence-transformers
faiss-cpu
tqdm
httpx
python-magic
python-multipart
